+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Fix Scroll and Scrollspy Nav components handle same site links only
+- Fix Drop positioning within not scrollable viewport
 
 ## 3.15.14 (November 21, 2022)