+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Fix Scroll and Scrollspy Nav components handle same site links only
 - Fix Drop positioning within not scrollable viewport
+- Fix overscroll behavior in Dropbar
 
 ## 3.15.14 (November 21, 2022)